refactor pkg/testutils package
