+++ tests/e2e/commontestutils/logs.go
@@ -1,4 +1,4 @@
-package testutils
+package commontestutils
 
 import (
 	"bytes"
