+++ tests/e2e/commontestutils/metrics.go
@@ -1,4 +1,4 @@
-package testutils
+package commontestutils
 
 import (
 	"context"
