+++ tests/e2e/commontestutils/rbac.go
@@ -1,4 +1,4 @@
-package testutils
+package commontestutils
 
 import (
 	"context"
