+++ tests/e2e/commontestutils/statefulset.go
@@ -1,4 +1,4 @@
-package testutils
+package commontestutils
 
 import (
 	"context"
@@ -8,6 +8,7 @@ import (
 	apiappsv1 "k8s.io/api/apps/v1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
+	"github.com/kyma-project/lifecycle-manager/pkg/testutils"
 	"github.com/kyma-project/lifecycle-manager/pkg/util"
 )
 
@@ -17,7 +18,7 @@ func StatefulSetIsReady(ctx context.Context, clnt client.Client, name, namespace
 	statefulSet, err := GetStatefulSet(ctx, clnt, name, namespace)
 	if err != nil {
 		if util.IsNotFound(err) {
-			return ErrNotFound
+			return testutils.ErrNotFound
 		}
 		return fmt.Errorf("could not get statefulset: %w", err)
 	}
