+++ tests/e2e/commontestutils/watcher.go
@@ -1,4 +1,4 @@
-package testutils
+package commontestutils
 
 import (
 	"bytes"
