+++ tests/e2e/mandatory_modules_with_old_naming_pattern_metrics_test.go
@@ -11,6 +11,7 @@ import (
 	. "github.com/onsi/gomega"
 
 	. "github.com/kyma-project/lifecycle-manager/pkg/testutils"
+	. "github.com/kyma-project/lifecycle-manager/tests/e2e/commontestutils"
 )
 
 var _ = Describe("Mandatory Module With Old Naming Pattern Metrics", Ordered, func() {
