+++ tests/e2e/module_consistency_test.go
@@ -11,6 +11,7 @@ import (
 	"github.com/kyma-project/lifecycle-manager/api/v1beta2"
 
 	. "github.com/kyma-project/lifecycle-manager/pkg/testutils"
+	. "github.com/kyma-project/lifecycle-manager/tests/e2e/commontestutils"
 )
 
 var _ = Describe("Module Keep Consistent After Deploy", Ordered, func() {
