+++ tests/e2e/modulereleasemeta_module_upgrade_new_version_test.go
@@ -7,6 +7,7 @@ import (
 	"github.com/kyma-project/lifecycle-manager/api/shared"
 	"github.com/kyma-project/lifecycle-manager/api/v1beta2"
 	. "github.com/kyma-project/lifecycle-manager/pkg/testutils"
+	. "github.com/kyma-project/lifecycle-manager/tests/e2e/commontestutils"
 )
 
 var _ = Describe("Module with ModuleReleaseMeta Upgrade By New Version", Ordered, func() {
